+++ CHANGELOG.md
@@ -106,6 +106,8 @@
     * `stubbles\predicate\IsMailAddress`
     * `stubbles\predicate\IsOneOf`
     * `stubbles\predicate\Regex`
+  * `stubbles\lang\reflect()` can now also reflect functions
+  * `stubbles\lang\reflect()` now throws a `stubbles\lang\exception\IllegalArgumentException` if the value can not be reflected
 
 
 3.5.3 (2014-05-07)